+++ src/services/message/message.test.ts
@@ -1,5 +1,5 @@
 import { setupRecorder } from "nock-record";
-import { AddChatItemAction, Masterchat, timeoutThen } from "../../..";
+import { AddChatItemAction, Masterchat, timeoutThen } from "../..";
 
 const id = process.env.MC_MSG_TEST_ID;
 const channelId = process.env.MC_MSG_TEST_CHANNEL_ID;