+++ examples/manager.ts
@@ -1,0 +1,26 @@
+import fetch from "cross-fetch";
+import { MasterchatManager } from "masterchat";
+
+async function getStreams() {
+  const res = await fetch(
+    "https://holodex.net/api/v2/live?org=Hololive&max_upcoming_hours=0"
+  );
+  const streams = await res.json();
+  return streams.slice(0, 3);
+}
+
+async function main() {
+  const manager = new MasterchatManager({ isLive: true });
+
+  manager.on("chats", (videoId, chats) =>
+    console.log(videoId, "received", chats.length, "chats")
+  );
+  manager.on("end", (videoId) => console.log(videoId, "ended"));
+  manager.on("error", (videoId, err) => console.error(videoId, err.message));
+
+  for (const stream of await getStreams()) {
+    manager.subscribe(stream.id, stream.channel.id);
+  }
+}
+
+main();