+++ CHANGELOG.md
@@ -2,11 +2,14 @@
 
 ## Unreleased
 
+- Masterchat agent/manager for handling multiple instances of Masterchat
+- now MC requires Node.js v16.6 or higher
 - re-export more yt types
 
 ### utils
 
-- `normalizeVideoId` will returns `undefined` when the given string doesn't contains any valid id pattern
+- Renamed `normalizedVideoId` -> `toVideoId`
+- `toVideoId` will returns `undefined` when the given string doesn't contains any valid id pattern
 - fix format issue of default emoji handler of `runsToString`
 
 ## v0.10.0